Extract `emitDirIndexShims` so the shim-collision test imports it directly instead of spawning `bun`
`scripts/build.ts` statically imports `bun`, so it can't be imported under the Node vitest runner; the collision-guard test worked around this by driving the emitter through a generated `drive.mjs` in a `bun` child process. Move `emitDirIndexShims` (and its `moduleHasDefaultExport` helper) into `scripts/emit-dir-index-shims.ts`, depending only on `node:fs`, and have both `build.ts` and the test import it. The test collapses to a direct `rejects.toThrow` assertion. Emitted shim bytes are unchanged; the require-argument is assembled indirectly so a CommonJS-interop scanner doesn't resolve the shim's literal `require('…')`.
