+++ drizzle-orm/scripts/build.ts
@@ -1,9 +1,10 @@
 #!/usr/bin/env bun
 import { $ } from 'bun';
 import { globSync } from 'glob';
-import { existsSync, mkdirSync, renameSync, rmSync } from 'node:fs';
+import { mkdirSync, renameSync, rmSync } from 'node:fs';
 import * as fs from 'node:fs/promises';
 import { build as tsdown } from 'tsdown';
+import { emitDirIndexShims } from './emit-dir-index-shims.ts';
 
 const entries = globSync('src/**/*.ts', { ignore: ['src/**/*.test.ts'] });
 
@@ -13,45 +14,6 @@ async function copyPackageJson() {
 	// shims are the only dynamic piece — see emitDirIndexShims.
 	const pkg = JSON.parse(await fs.readFile('package.json', 'utf8'));
 	await fs.writeFile('dist.new/package.json', JSON.stringify(pkg, null, 2));
-}
-
-async function moduleHasDefaultExport(modulePath: string): Promise<boolean> {
-	if (!existsSync(modulePath)) return false;
-	const src = await fs.readFile(modulePath, 'utf8');
-	return /(^|\n)\s*export\s+default\b/.test(src) || /(^|\n)\s*export\s*\{[^}]*\bdefault\b[^}]*\}/.test(src);
-}
-
-export async function emitDirIndexShims(entries: string[], outDir: string) {
-	const dirIndex = entries
-		.map((raw) => raw.match(/src\/(.*)\.ts/)![1]!)
-		.filter((e) => e.endsWith('/index') && e !== 'index')
-		.map((e) => e.replace(/\/index$/, ''));
-
-	for (const x of dirIndex) {
-		const target = `${outDir}/${x}.js`;
-		// A real tsdown-emitted artifact already at this path means a sibling leaf
-		// src/${x}.ts collides with src/${x}/index.ts; shadowing it would ship a
-		// broken or wrong subpath, so fail the build loudly instead.
-		if (existsSync(target)) {
-			throw new Error(
-				`refusing to overwrite source-emitted artifact: ${target} with a directory-index shim — `
-					+ `src/${x}.ts collides with src/${x}/index.ts. Resolve the collision at the source.`,
-			);
-		}
-
-		// The shim sits beside the `x` directory (dist/<x>.js next to dist/<x>/), so its
-		// re-export target is relative to the shim's OWN dir — use the basename, not the full
-		// nested path, or a nested shim like dist/pg-core/async.js points at the nonexistent
-		// dist/pg-core/pg-core/async/index.js.
-		const baseName = x.slice(x.lastIndexOf('/') + 1);
-		const hasDefault = await moduleHasDefaultExport(`${outDir}/${x}/index.js`);
-		const dflt = hasDefault ? `\nexport { default } from './${baseName}/index.js';` : '';
-
-		await fs.writeFile(target, `export * from './${baseName}/index.js';${dflt}`);
-		await fs.writeFile(`${outDir}/${x}.cjs`, `module.exports = require('./${baseName}/index.cjs');`);
-		await fs.writeFile(`${outDir}/${x}.d.ts`, `export * from './${baseName}/index.js';${dflt}`);
-		await fs.writeFile(`${outDir}/${x}.d.cts`, `export * from './${baseName}/index.cjs';`);
-	}
 }
 
 async function main() {
